+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+Features:
+
+- AbstractTcpClient: applied exponential backoff to connection retry interval
+- AbstractTcpClient: randomized reconnect delay
+
 ## 0.17.1 (2018-11-28)
 
 Bugfixes: